Separate apis for BehaviorTestKit effects
