+++ akka-actor-typed-tests/src/test/java/jdocs/akka/typed/testing/sync/BasicSyncTestingTest.java
@@ -3,7 +3,6 @@ package jdocs.akka.typed.testing.sync;
 //#imports
 import akka.actor.typed.*;
 import akka.actor.typed.javadsl.*;
-import akka.testkit.typed.*;
 import akka.testkit.typed.javadsl.*;
 //#imports
 import org.junit.Test;
@@ -69,7 +68,7 @@ public class BasicSyncTestingTest extends JUnitSuite {
     //#test-child
     BehaviorTestKit<Command> test = BehaviorTestKit.create(myBehavior);
     test.run(new CreateAChild("child"));
-    test.expectEffect(new Effect.Spawned(childActor, "child", Props.empty()));
+    test.expectEffect(Effects.spawned(childActor, "child", Props.empty()));
     //#test-child
   }
 
@@ -78,7 +77,7 @@ public class BasicSyncTestingTest extends JUnitSuite {
     //#test-anonymous-child
     BehaviorTestKit<Command> test = BehaviorTestKit.create(myBehavior);
     test.run(new CreateAnAnonymousChild());
-    test.expectEffect(new Effect.SpawnedAnonymous(childActor, Props.empty()));
+    test.expectEffect(Effects.spawnedAnonymous(childActor, Props.empty()));
     //#test-anonymous-child
   }
